Remove compiler synthesized struct field. (#230)
This was originally meant to contain per struct object verification information, such as the AbilitySet. With experience and knowledge gained with Move, this information won't be needed per object. Abilities are part of the type, known statically, and invariant. They could be stored with the per-type StructTypeInfo, or with other generic type and verification information that the runtime/VM will eventually use. We'll defer that decision until wider discussions around the future structure of the VM/runtime verification will work.